Move shared Updater logic into UpdaterCommon
