+++ Source/Core/UpdaterCommon/CMakeLists.txt
@@ -1,0 +1,8 @@
+add_library(updatercommon
+  UpdaterCommon.cpp)
+  
+target_link_libraries(updatercommon PRIVATE
+  uicommon
+  mbedtls
+  z
+  ed25519)
